Move alert handling to separate file
This is to cut down on the size of the file as more actions are added
